feat: Dynamic Price cards (server component)
